Project Files/requirements.txt
fastapi
uvicorn
python-dotenv
transformers
torch
pinecone
